requirements.txt
streamlit

opencv-python-headless

numpy

easyocr

ultralytics

pillow

pandas